ABI: add `AsyncActionCompletedHandler`
This adds the `AsyncActionCompletedHandler` type which provides the
means for implementing a concrete `IAsyncActionCompletedHandler`.

Due to the construction of `IUnknown(pUnk:)` normally being consuming,
we cannot derive from the interface definition.  However, as a
workaround, the `Interface` function provides the mechanism for
constructing the interface pointer from the type.  This allows us to
treat the concrete type as an implementation of the interface.