Support: add `Future` protocol
The `Future` protocol allows a conforming type to be used as a future.
The protocol is meant to be adopted by promises.  A conforming promise
provides the `get` method which will evaluate the promise and return a
value for the promise.  This is meant to be used in the implementation
for async routines.