ABI: extend the `IDispatcherQueueController` wrapping
This provides the first of the wrapping routines that truly bridge WinRT
to Swift.  The `ShutdownQueueAsync` function is further extended towards
a Swift friendly interface.  The `Async` suffix is stripped and the func
implemented as an `async` overload.  This function can now be invoked in
an async context with the `await` keyword enabling native concurrency to
be used with the WinRT model.